pyproject.toml
[tool.poetry]
name = "agents"
version = "0.1.0"
description = ""
authors = ["HrithikMJ <hrithikj@anuntatech.com>"]
readme = "README.md"
package-mode = false 

[tool.poetry.dependencies]
python = "^3.12"
langchain-core = "^1.2.7"
langchain-cohere = "^0.5.0"
dotenv = "^0.9.9"
langchain-openai = "^1.1.7"
sentence-transformers = "^5.2.2"
langchain-huggingface = "^1.2.0"
ipykernel = "^7.1.0"
langchain-chroma = "^1.1.0"
langgraph = "^1.0.7"
langchain = "^1.2.8"
deepagents = "^0.3.9"
tavily-python = "^0.7.21"
structlog = "^25.5.0"
loguru = "^0.7.3"
python-telegram-bot = "^22.6"
telegramify-markdown = "^0.5.4"
langgraph-checkpoint-postgres = "^3.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
